+++ app.py
@@ -4,12 +4,15 @@ from authlib.integrations.flask_client import OAuth
 from models import connect_to_db, db, User
 
 import os
-import bcrypt
+
+from login.login_routes import login_blueprint
 
 #running an instance of flask
 app = Flask(__name__)
 app.secret_key = os.environ['APP_SECRET_KEY']
 oauth = OAuth(app)
+app.register_blueprint(login_blueprint)
+
 
 # register authentication
 google = oauth.register(
@@ -25,7 +28,7 @@ google = oauth.register(
 )
 
 # below is registering a remote app for twitter; twitter is an OAuth 1.0 service
-# request_token_url is used to detect if client is OAuth 1 or OAuth 2 cient.
+# request_token_url is used to detect if client is OAuth 1 or OAuth 2 client.
 twitter = oauth.register(
     name = 'twitter',
     client_id = os.environ['TWITTER_CLIENT_ID'],
@@ -148,74 +151,12 @@ def authorize_twitter():
 #     return redirect('/')
 
 
-@app.route('/logout')
-def logout():
-    """logs out any user"""
-
-    for key in list(session.keys()):
-        session.pop(key)
-    return redirect('/')
-
-
-@app.route('/sign-up', methods=['POST'])
-def signup():
-    """manual user registration; redirects user to their profile page"""
-
-    first_name = request.form.get('first_name')
-    email = request.form.get('email')
-    password = request.form.get('password')
-    hashed_password = bcrypt.hashpw(password.encode("utf-8"), bcrypt.gensalt())
-
-    # TODO add regex to form to ensure email and password fit parameters
-    # error messages if password does not meet expectations?
-    if hashed_password:
-        new_user = User(first_name=first_name, email=email, password=hashed_password)
-    else:
-        return f"Please enter a valid password"
-
-    # add to db here
-    add_user_db(new_user)
-
-    # adding session
-    user_id = new_user.user_id
-    session["user_id"] = user_id
-
-    return redirect(f'/user/{user_id}')
-
-
 def add_user_db(new_user):
     """adds user to db"""
 
     db.session.add(new_user)
     db.session.commit()
     pass
-
-@app.route('/login', methods=['POST'])
-def login():
-    """logging in existing users who manually registered on app"""
-
-    # get user email from form
-    email = request.form.get('email')
-    password = (request.form.get('password')).encode('utf-8')
-
-    # search db for user
-    existing_user = User.query.filter_by(email=email).first()
-    existing_password = existing_user.password
-    # user_email = existing_user.email
-    user_id = existing_user.user_id
-
-    if not existing_user:
-        flash('No user found with this email. Please try again')
-        return redirect('/')
-
-    # compare password entered with password from db
-    if bcrypt.checkpw(password, existing_password):
-        # flash(f'Welcome back, {user_email}!')
-        session["user_id"] = user_id
-        return redirect(f'/user/{user_id}')
-    else:
-        flash('Password does not match. Please try again.')
-        return redirect('/')
 
 
 @app.route('/user/<int:user_id>')
@@ -224,6 +165,8 @@ def user_profile(user_id):
     # for now, only using user_id
     # with future functionality for saving herbs, will query db for that
     # and return onto page; one to many relationship
+    # TODO: MAKE SURE USER IS LOGGED IN (another fxn?), DONT JUST QUERY FOR USER
+    # todo: use .get_or_404()
 
     first_name = (User.query.filter_by(user_id=user_id).first()).first_name
     return render_template('profile-page.html', first_name=first_name)
@@ -232,9 +175,9 @@ def user_profile(user_id):
 #############
 # TODO check if i need to update token for google. search 'refresh_token' in docs
 # TODO: utilize ajax for google/twitter/fb buttons on homepage
-# TODO: add in module for each registry/route for fb/google/twitter -- this
-# TODO: can consolidate how many routes i need to have in app; have a login module?
-# TODO: create a function for adding a user to db (im reusing db.add/db.commit in routes)
+# TODO: add in module for each registry/route for fb/google/twitter
+# TODO: add blueprints to consolidate how many routes i need to have in app; have a login module?
+# TODO: have a delete function for user to delete their info from db
 #############
